// src/app/open-api/generated/model/email-setting.ts
/**
 * GED API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SettingBase } from './setting-base';
import { FlowType } from './flow-type';
import { EmailAccessProtocol } from './email-access-protocol';


export interface EmailSetting extends SettingBase { 
    host: string;
    port: string;
    username: string;
    password: string;
    protocol: EmailAccessProtocol;
}
export namespace EmailSetting {
}